Move Primitive.AddressDiscovery namespace in Address.Discovery .
