+++ lib/wallet/api/http/Cardano/Wallet/Api/Link.hs
@@ -125,6 +125,8 @@ import Prelude
 
 import Cardano.Wallet.Address.Derivation
     ( DerivationIndex, Role )
+import Cardano.Wallet.Address.Discovery.Shared
+    ( CredentialType (..) )
 import Cardano.Wallet.Api.Types
     ( ApiAddressInspectData (..)
     , ApiPoolSpecifier
@@ -139,8 +141,6 @@ import Cardano.Wallet.Api.Types.SchemaMetadata
     ( TxMetadataSchema (..), toSimpleMetadataFlag )
 import Cardano.Wallet.Api.Types.Transaction
     ( ApiLimit )
-import Cardano.Wallet.Primitive.AddressDiscovery.Shared
-    ( CredentialType (..) )
 import Cardano.Wallet.Primitive.Types
     ( SmashServer, SortOrder, WalletId (..) )
 import Cardano.Wallet.Primitive.Types.Address
