+++ lib/wallet/api/http/Cardano/Wallet/Api/Types/Key.hs
@@ -36,10 +36,10 @@ import Cardano.Wallet.Address.Derivation
     ( Depth (..), DerivationType (..), Index (..), Role (..) )
 import Cardano.Wallet.Address.Derivation.SharedKey
     ( purposeCIP1854 )
+import Cardano.Wallet.Address.Discovery.Sequential
+    ( purposeCIP1852 )
 import Cardano.Wallet.Api.Lib.Options
     ( DefaultSum (..) )
-import Cardano.Wallet.Primitive.AddressDiscovery.Sequential
-    ( purposeCIP1852 )
 import Codec.Binary.Bech32
     ( dataPartFromBytes, dataPartToBytes )
 import Codec.Binary.Bech32.TH
