+++ lib/wallet/integration/src/Test/Integration/Scenario/API/Byron/HWWallets.hs
@@ -23,6 +23,8 @@ import Cardano.Mnemonic
     )
 import Cardano.Wallet.Address.Derivation
     ( HardDerivation (..), PersistPublicKey (..), WalletKey (..) )
+import Cardano.Wallet.Address.Discovery.Sequential
+    ( defaultAddressPoolGap, getAddressPoolGap )
 import Cardano.Wallet.Api.Types
     ( AddressAmount (..)
     , ApiAddress (..)
@@ -34,8 +36,6 @@ import Cardano.Wallet.Api.Types
     , ApiUtxoStatistics
     , WalletStyle (..)
     )
-import Cardano.Wallet.Primitive.AddressDiscovery.Sequential
-    ( defaultAddressPoolGap, getAddressPoolGap )
 import Cardano.Wallet.Primitive.Types.Address
     ( AddressState (..) )
 import Cardano.Wallet.Read.NetworkId
