+++ lib/wallet/integration/src/Test/Integration/Scenario/API/Shared/Addresses.hs
@@ -21,10 +21,10 @@ import Prelude
 
 import Cardano.Wallet.Address.Derivation.SharedKey
     ( purposeCIP1854 )
+import Cardano.Wallet.Address.Discovery.Sequential
+    ( defaultAddressPoolGap, getAddressPoolGap )
 import Cardano.Wallet.Api.Types
     ( ApiAddressWithPath, ApiSharedWallet (..), WalletStyle (..) )
-import Cardano.Wallet.Primitive.AddressDiscovery.Sequential
-    ( defaultAddressPoolGap, getAddressPoolGap )
 import Cardano.Wallet.Primitive.Types.Address
     ( AddressState (..) )
 import Cardano.Wallet.Read.NetworkId
